requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
sse-starlette==2.1.3
pytest==8.3.3
httpx==0.27.2
pytest-asyncio==0.24.0
anthropic==0.42.0
python-dotenv==1.1.0
